make branches in to a seperate entity and change existing pharmacy entity

+++ src/main/java/com/example/demo/controller/PharmacyController.java
@@ -18,6 +18,7 @@ public class PharmacyController {
     public ResponseEntity<Pharmacy> create(@RequestBody Pharmacy pharmacy){
         return ResponseEntity.ok(service.createPharmacy(pharmacy));
     }
+    
     @PostMapping("/createpharmacies")
     public ResponseEntity<List<Pharmacy>> createPharmacies(@RequestBody List<Pharmacy> pharmacies){
         return ResponseEntity.ok(service.createPharmacies(pharmacies));
@@ -29,9 +30,7 @@ public class PharmacyController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Pharmacy> getPharmacyById(@PathVariable int id){
+    public ResponseEntity<Pharmacy> getPharmacyById(@PathVariable Long id){
         return ResponseEntity.ok(service.getPharmacyById(id));
     }
-
-
 }
