+++ src/main/java/com/example/demo/repository/PharmacyRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Pharmacy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PharmacyRepository extends JpaRepository<Pharmacy,Integer> {
+public interface PharmacyRepository extends JpaRepository<Pharmacy, Long> {
 }
